+++ WineManagement/BusinessLayer/Service/AuthService.cs
@@ -51,7 +51,7 @@ namespace BusinessLayer.Service
 							{
 								AccountId= account.AccountId,
 								Username = account.Username,
-								RoleId= account.RoleId,
+								RoleId= (int)account.RoleId,
 								Status= account.Status
 							},
 						},
@@ -70,7 +70,7 @@ namespace BusinessLayer.Service
 						{
 							AccountId = account.AccountId,
 							Username = account.Username,
-							RoleId = account.RoleId,
+							RoleId = (int)account.RoleId,
 							Status = account.Status
 						},
 					},
